+++ Product/src/main/java/org/example/product/controller/ProductController.java
@@ -21,7 +21,7 @@ public class ProductController {
     }
 
     @GetMapping("/{productId}")
-    public ProductDTO getProductById(@PathVariable Integer productId) {
+    public ProductDTO getProductById(@PathVariable Long productId) {
         return productService.getProductById(productId);
     }
